refactor: move all permissions functionalities under `permissions.ts`
